CryptoView/requirements.txt
dash==1.13.3
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.0
dash-table==4.8.1
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
numpy==1.19.0
pandas==1.0.5
plotly==4.8.1
psutil==5.7.0
python-binance==0.7.5
python-dateutil==2.8.1
python-dotenv==0.13.0
termcolor==1.1.0
watchdog==0.10.2
